pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.11.0"]
build-backend = "uv_build"

[project]
name = "danish_asr"
version = "0.0.1"
description = "Fine-tuning and Evaluation of Modern ASR Models for Danish"
authors = [
  { name = "Your Name", email = "you@example.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Deep Learning
    "torch>=2.7.0",
    "pytorch-lightning>=2.1.0",
    "torchmetrics>=1.0.0",
    "rich>=13.0.0",
    # MLOps Tools
    "wandb[workspaces]>=0.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "loguru>=0.7.3",
    # API & CLI
    "fastapi>=0.115.0",
    "typer>=0.9.0",
    "uvicorn>=0.30.0",
    # Data Science
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "pandas>=2.0.0",
    "python-multipart>=0.0.21",
    "httpx>=0.28.1",
    "psutil>=6.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "evidently>=0.7.20",
    "prometheus-client>=0.24.1",
    "scipy>=1.16.3",
]

[dependency-groups]
dev = [
    "coverage>=7.6.0",
    "invoke>=2.2.0",
    "mkdocs-material>=9.4.0",
    "mkdocs>=1.6.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "mkdocstrings-python>=1.12.0",
    "locust>=2.20.0",
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "bandit>=1.9.3",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "NPY",  # NumPy-specific rules
    "C4",   # flake8-comprehensions
    "RET",  # flake8-return
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["danish_asr"]
known-third-party = ["wandb", "torch", "lightning"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D103"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "requires_data: mark test as requiring actual data files (skipped in CI)",
    "slow: mark test as slow running",
]
addopts = "-v --tb=short"

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["wandb", "wandb.*"]
follow_imports = "skip"

[tool.coverage.run]
omit = [
    "tests/*",
    "*/config*.py",
    ".venv/*",
]

[project.scripts]
train = "danish_asr.train:train"
sweep-train = "danish_asr.sweep_train:main"
evaluate = "danish_asr.evaluate:evaluate_model"
api = "danish_asr.api:app"
